Remove TaintedMap interface
