+++ dd-java-agent/agent-iast/src/jmh/java/com/datadog/iast/taint/TaintedMapPutsBenchmark.java
@@ -31,7 +31,7 @@ import org.openjdk.jmh.infra.Blackhole;
 @State(Scope.Benchmark)
 public class TaintedMapPutsBenchmark {
 
-  private static final int INITIAL_OP_COUNT = DefaultTaintedMap.DEFAULT_FLAT_MODE_THRESHOLD;
+  private static final int INITIAL_OP_COUNT = TaintedMap.DEFAULT_FLAT_MODE_THRESHOLD;
   private static final int OP_COUNT = 1024;
 
   private static final Range[] EMPTY_RANGES = new Range[0];
@@ -42,7 +42,7 @@ public class TaintedMapPutsBenchmark {
 
   @Setup(Level.Iteration)
   public void setup() {
-    map = new DefaultTaintedMap();
+    map = new TaintedMap();
     objectBuffer = new CircularBuffer<>(OP_COUNT);
     initialObjectList = new ArrayList<>(INITIAL_OP_COUNT);
     for (int i = 0; i < INITIAL_OP_COUNT; i++) {
